Modify TSP formulation.
Firstly, the TSP solver needs pairwise distances to be calculated.
These form a symmetric matrix, so half the calculations can be removed.

Instead of using one single distance calculation, we use the Euclidean
heuristic (straight-line distance) if there is line-of-sight between two
points. However, if there is no line of sight, we will use the RRT(*)
  distance. (a path is calculated).

Since we already calculated a path to obtain the distance, this path is
re-used when the TSP problem is solved.